Reorganize functions and implement help functions
