+++ WooCommerce/src/main/AndroidManifest.xml
@@ -64,7 +64,7 @@
         </activity>
 
         <activity
-            android:name=".PostLoginSiteDiscoveryActivity"
+            android:name=".ui.sitediscovery.PostLoginSiteDiscoveryActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
